Correção, utilizando laravel no backend

+++ imobiliaria-api/config/jwt.php
@@ -9,7 +9,7 @@ return [
         'passphrase' => env('JWT_PASSPHRASE'),
     ],
 
-    'ttl' => env('JWT_TTL', 60),
+    'ttl' => (int) env('JWT_TTL', 60),
     'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
     'algo' => env('JWT_ALGO', 'HS256'),
     'required_claims' => ['iss', 'iat', 'exp', 'nbf', 'sub', 'jti'],
